Extract e2e qa databases helpers
