+++ frontend/test/__support__/e2e/helpers/e2e-qa-databases-helpers.js
@@ -1,0 +1,67 @@
+/*****************************************
+ **            QA DATABASES             **
+ ******************************************/
+
+export function addMongoDatabase(name = "QA Mongo4") {
+  // https://hub.docker.com/layers/metabase/qa-databases/mongo-sample-4.0/images/sha256-3f568127248b6c6dba0b114b65dc3b3bf69bf4c804310eb57b4e3de6eda989cf
+  addQADatabase("mongo", name, 27017);
+}
+
+export function addPostgresDatabase(name = "QA Postgres12") {
+  // https://hub.docker.com/layers/metabase/qa-databases/postgres-sample-12/images/sha256-80bbef27dc52552d6dc64b52796ba356d7541e7bba172740336d7b8a64859cf8
+  addQADatabase("postgres", name, 5432);
+}
+
+export function addMySQLDatabase(name = "QA MySQL8") {
+  // https://hub.docker.com/layers/metabase/qa-databases/mysql-sample-8/images/sha256-df67db50379ec59ac3a437b5205871f85ab519ce8d2cdc526e9313354d00f9d4
+  addQADatabase("mysql", name, 3306);
+}
+
+function addQADatabase(engine, db_display_name, port) {
+  const PASS_KEY = engine === "mongo" ? "pass" : "password";
+  const AUTH_DB = engine === "mongo" ? "admin" : null;
+  const OPTIONS = engine === "mysql" ? "allowPublicKeyRetrieval=true" : null;
+
+  cy.log(`**-- Adding ${engine.toUpperCase()} DB --**`);
+  cy.request("POST", "/api/database", {
+    engine: engine,
+    name: db_display_name,
+    details: {
+      dbname: "sample",
+      host: "localhost",
+      port: port,
+      user: "metabase",
+      [PASS_KEY]: "metasample123", // NOTE: we're inconsistent in where we use `pass` vs `password` as a key
+      authdb: AUTH_DB,
+      "additional-options": OPTIONS,
+      "use-srv": false,
+      "tunnel-enabled": false,
+    },
+    auto_run_queries: true,
+    is_full_sync: true,
+    schedules: {
+      cache_field_values: {
+        schedule_day: null,
+        schedule_frame: null,
+        schedule_hour: 0,
+        schedule_type: "daily",
+      },
+      metadata_sync: {
+        schedule_day: null,
+        schedule_frame: null,
+        schedule_hour: null,
+        schedule_type: "hourly",
+      },
+    },
+  }).then(({ status }) => {
+    expect(status).to.equal(200);
+  });
+
+  // Make sure we have all the metadata because we'll need to use it in tests
+  cy.request("POST", "/api/database/2/sync_schema").then(({ status }) => {
+    expect(status).to.equal(200);
+  });
+  cy.request("POST", "/api/database/2/rescan_values").then(({ status }) => {
+    expect(status).to.equal(200);
+  });
+}
